+++ src/components/Ui/zoomInImg.jsx
@@ -11,7 +11,7 @@ const ImageModal = ({ src, alt, onClose }) => {
   return (
     <div className={styles.modal} onClick={onClose} role="button" tabIndex={0} onKeyDown={handleKeyDown}>
       <div className={styles.modalContent}>
-        <img src={src} alt={alt} className={styles.zoomInImg} />
+        <img src={src} alt={alt} className={styles.zoomInImg} loading='lazy'/>
       </div>
     </div>
   )
